refactor: organize transpiler directories by responsibility
- Move TranspilerState.ts from types/ to state/ (stateful class, not type definition)
- Move BITMAP_SIZE.ts and BITMAP_BACKING_TYPE.ts from types/ to constants/ (runtime lookup tables)
- Update imports in Transpiler.ts and BitmapCollector.ts
- Move TranspilerState.test.ts to state/__tests__/
- Update CLAUDE.md to document the new directory structure

This aligns with the 4-layer architecture:
- types/ for type definitions only (interfaces, type aliases, enums)
- state/ for stateful classes (CodeGenState, SymbolRegistry, TranspilerState)
- constants/ for runtime lookup tables
