+++ src/transpiler/constants/BITMAP_BACKING_TYPE.ts
