+++ src/transpiler/constants/BITMAP_SIZE.ts
